+++ lib/src/builders/builders.dart
@@ -1,2 +1,4 @@
 export 'select/select_builder.dart';
 export 'insert/insert_builder.dart';
+export 'update/update_builder.dart';
+export 'delete/delete_builder.dart';