+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## next
+* Adds `delete` method.
+
 ## 0.2.2
 
 * Adds `TypeAdapter`.